+++ recipe/thirdpartyemailpassword/tpepmodels/apiInterface.go
@@ -27,7 +27,7 @@ type APIInterface struct {
 	AppleRedirectHandlerPOST       *func(code string, state string, options tpmodels.APIOptions, userContext supertokens.UserContext) error
 	EmailPasswordEmailExistsGET    *func(email string, options epmodels.APIOptions, userContext supertokens.UserContext) (epmodels.EmailExistsGETResponse, error)
 	GeneratePasswordResetTokenPOST *func(formFields []epmodels.TypeFormField, options epmodels.APIOptions, userContext supertokens.UserContext) (epmodels.GeneratePasswordResetTokenPOSTResponse, error)
-	PasswordResetPOST              *func(formFields []epmodels.TypeFormField, token string, options epmodels.APIOptions, userContext supertokens.UserContext) (epmodels.ResetPasswordUsingTokenResponse, error)
+	PasswordResetPOST              *func(formFields []epmodels.TypeFormField, token string, options epmodels.APIOptions, userContext supertokens.UserContext) (epmodels.ResetPasswordPOSTResponse, error)
 	ThirdPartySignInUpPOST         *func(provider tpmodels.TypeProvider, code string, authCodeResponse interface{}, redirectURI string, options tpmodels.APIOptions, userContext supertokens.UserContext) (ThirdPartyOutput, error)
 	EmailPasswordSignInPOST        *func(formFields []epmodels.TypeFormField, options epmodels.APIOptions, userContext supertokens.UserContext) (SignInPOSTResponse, error)
 	EmailPasswordSignUpPOST        *func(formFields []epmodels.TypeFormField, options epmodels.APIOptions, userContext supertokens.UserContext) (SignUpPOSTResponse, error)
@@ -39,6 +39,7 @@ type SignUpPOSTResponse struct {
 		Session sessmodels.SessionContainer
 	}
 	EmailAlreadyExistsError *struct{}
+	GeneralError            *supertokens.GeneralErrorResponse
 }
 
 type SignInPOSTResponse struct {
@@ -47,6 +48,7 @@ type SignInPOSTResponse struct {
 		Session sessmodels.SessionContainer
 	}
 	WrongCredentialsError *struct{}
+	GeneralError          *supertokens.GeneralErrorResponse
 }
 
 type EmailpasswordInput struct {